src: initial refactoring
