+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "hid-tminit"]
-	path = hid-tminit
-	url = https://github.com/scarburato/hid-tminit.git
+[submodule "src/hid-tminit"]
+	path = src/hid-tminit
+	url = https://github.com/scarburato/hid-tminit
