--- Kbuild
@@ -1,2 +1,0 @@
-obj-m := hid-tmff-new.o
-hid-tmff-new-y := hid-tmff2.o hid-tmt300rs.o hid-tmt248.o hid-tmtx.o
