+++ docs/FFBEffects-T248.md
@@ -1,0 +1,7 @@
+# Force Feedback Effects - Thrustmaster T248
+
+## Open: 
+Identical to T300RS
+
+## Close:
+Identical to T300RS
