+++ docs/TODO.md
@@ -1,0 +1,10 @@
+# TO-DO list
+
+- Make code more general to allow for stuff like T150/T248/T500 to be relatively
+easily included.
+
+- Use workqueues instead of doing everything in t300rs_timer
+
+### T248
+    - open/close seems to crash the wheel and the kernel along with it, check how Windows handles it
+    - fix rdesc, 0x0a somewhere needs to be set to 0x60
