--- ps4-input.txt
@@ -1,54 +1,0 @@
-1 ff00.0021 = 0
-2 ff00.0021 = 0
-3 ff00.0021 = 0
-4 ff00.0021 = 0
-5 ff00.0021 = 0
-6 ff00.0021 = 0
-7 ff00.0021 = 0
-8 ff00.0021 = 0
-9 ff00.0021 = 0
-10 ff00.0021 = 0
-11 ff00.0021 = 0
-12 ff00.0021 = 0
-13 ff00.0021 = 0
-14 ff00.0021 = 0
-15 ff00.0021 = 0
-16 ff00.0021 = 0
-17 ff00.0021 = 0
-18 ff00.0021 = 0
-19 ff00.0021 = 0
-20 ff00.0021 = 0
-21 ff00.0021 = 0
-22 ff00.0021 = 0
-23 ff00.0021 = 0
-24 ff00.0021 = 0
-25 ff00.0021 = 0
-26 ff00.0021 = 0
-27 ff00.0021 = 0
-28 ff00.0021 = 0
-29 ff00.0021 = 0
-30 ff00.0021 = 0
-31 ff00.0021 = 0
-32 ff00.0021 = 0
-33 ff00.0021 = 0
-34 ff00.0021 = 226 /* wheel */
-35 ff00.0021 = 125 /* wheel */
-36 ff00.0021 = 255 /* gas */
-37 ff00.0021 = 255 /* gas */
-38 ff00.0021 = 255 /* brake */
-39 ff00.0021 = 255 /* brake */
-40 ff00.0021 = 255 /* clutch */
-41 ff00.0021 = 255 /* clutch */
-42 ff00.0021 = 0
-43 ff00.0021 = 255
-44 ff00.0021 = 255
-45 ff00.0021 = 0
-46 ff00.0021 = 0
-47 ff00.0021 = 0
-48 ff00.0021 = 0
-49 ff00.0021 = 0
-50 ff00.0021 = 0
-51 ff00.0021 = 0
-52 ff00.0021 = 0
-53 ff00.0021 = 0
-54 ff00.0021 = 0
