+++ src/Kbuild
@@ -1,0 +1,2 @@
+obj-m := hid-tmff-new.o
+hid-tmff-new-y := hid-tmff2.o hid-tmt300rs/hid-tmt300rs.o hid-tmt248/hid-tmt248.o hid-tmtx/hid-tmtx.o
