+++ src/hid-tmt300rs/hid-tmt300rs.h
@@ -2,8 +2,8 @@
 #ifndef __HID_TMT300RS_H
 #define __HID_TMT300RS_H
 
-#include "hid-tmff2.h"
+#include "../hid-tmff2.h"
 
 int t300rs_populate_api(struct tmff2_device_entry *tmff2);
 
-#endif /* __HID_TM300RS_H */
+#endif
