--- t248-effects.txt
@@ -1,2 +1,0 @@
-Open: Identical to T300
-Close: Identical to T300
